refactor(sponsor): consolidate SponsorType definition and remove unused type file
