+++ frontend/admin/app/(app)/sponsor/_components/SponsorTable.tsx
@@ -14,7 +14,7 @@ import {
 } from "@heroui/react";
 import { EllipsisVertical, Pen, Trash } from "lucide-react";
 import type { Sponsors } from "@/types/sponsors";
-import type { SponsorType } from "@/types/sponsors-type";
+import type { SponsorType } from "@/types/sponsors";
 import { SponsorModal } from "./SponsorModal";
 
 interface SponsorTableProps {
