--- frontend/admin/types/sponsors-type.d.ts
@@ -1,4 +1,0 @@
-export type SponsorType = {
-    _id: string;
-    name: string;
-}
